Extract OutputChain
