+++ lib/fluent/load.rb
@@ -32,6 +32,7 @@ require 'fluent/formatter'
 require 'fluent/event'
 require 'fluent/buffer'
 require 'fluent/input'
+require 'fluent/output_chain'
 require 'fluent/output'
 require 'fluent/filter'
 require 'fluent/match'
