+++ lib/fluent/output.rb
@@ -20,50 +20,11 @@ require 'fluent/config'
 require 'fluent/configurable'
 require 'fluent/engine'
 require 'fluent/log'
+require 'fluent/output_chain'
 require 'fluent/plugin'
 require 'fluent/timezone'
 
 module Fluent
-  class OutputChain
-    def initialize(array, tag, es, chain=NullOutputChain.instance)
-      @array = array
-      @tag = tag
-      @es = es
-      @offset = 0
-      @chain = chain
-    end
-
-    def next
-      if @array.length <= @offset
-        return @chain.next
-      end
-      @offset += 1
-      result = @array[@offset-1].emit(@tag, @es, self)
-      result
-    end
-  end
-
-  class CopyOutputChain < OutputChain
-    def next
-      if @array.length <= @offset
-        return @chain.next
-      end
-      @offset += 1
-      es = @array.length > @offset ? @es.dup : @es
-      result = @array[@offset-1].emit(@tag, es, self)
-      result
-    end
-  end
-
-  class NullOutputChain
-    require 'singleton'
-    include Singleton
-
-    def next
-    end
-  end
-
-
   class Output
     include Configurable
     include PluginId
@@ -622,4 +583,3 @@ module Fluent
     #end
   end
 end
-
